[refactor] : event 패키지 구조 위치 변경
