+++ src/main/java/com/example/OMEB/domain/book/application/usecase/BookUseCase.java
@@ -8,11 +8,10 @@ import com.example.OMEB.domain.book.presentation.dto.request.BookApplicationRequ
 import com.example.OMEB.domain.book.presentation.dto.request.BookSearchRequest;
 import com.example.OMEB.domain.book.presentation.dto.response.BookInfoResponse;
 import com.example.OMEB.domain.book.presentation.dto.response.BookPageResponse;
-import com.example.OMEB.domain.book.presentation.dto.response.BookTitleInfoResponse;
 import com.example.OMEB.domain.book.presentation.dto.response.BookTitleListResponse;
 import com.example.OMEB.domain.book.presentation.dto.response.EmotionBookTitleInfoListResponse;
 import com.example.OMEB.domain.book.presentation.dto.response.NaverBookListResponse;
-import com.example.OMEB.domain.event.persistence.entity.EventView;
+import com.example.OMEB.global.event.persistence.entity.EventView;
 import com.example.OMEB.domain.review.persistence.vo.TagName;
 import com.example.OMEB.global.base.exception.ErrorCode;
 import com.example.OMEB.global.base.exception.ServiceException;
